Remove abstract OSGiTest class and dependency to CoreTest
This removes all dependencies to the to-be-removed JUnit 3-specific
CoreTest class. To this end, the OSGiTest class, which only contains
utility methods, is removed and the remaining utility fields and methods
moved to the existing OSGiTestActivator class.
